Heavy update
Improved significantly
Can now automatically construct database
Updates posts first seen time by monitoring the front page, so it has data on the post from before it gets locked. Additional method monitors the post after it is locked in the event that it is unlocked before leaving the front page.
new table 'blocked' can be used to suppress the bot from posting a given post to the subreddit even when it qualifies, allowing safe deletion of the surrogate post.