major update see notes
Modified to allow legaladvice posts to be inserted into db while still preventing crossposting.

Added try-excepts in some places to hopefully prevent crashes when reddit goes down. Can't really test it as it involves using the praw API which will fail if pointed at other sites regardless of connection status.

Added handler for cases when a post by [deleted] is on the frontpage. Should investigate to see if NULL overwrites username if user is deleted after the name is already recorded in DB.

Created bot capable of summarizing the statistical risk bot and posting to the subreddit. Ultimately probably not that useful but still nice tool.

Now capable of being auto-restarted with cron and flock.

statbot can now judge based on user.